+++ Plex/TaskScheduler.cs
@@ -12,54 +12,17 @@ namespace PlexDvrWaker.Plex
     /// </summary>
     internal class TaskScheduler
     {
-        private const string TASK_SCHEDULER_FOLDER = "Plex DVR Waker";
+        private const string TASK_SCHEDULER_FOLDER = Program.APP_FRIENDLY_NAME;
         private const string TASK_NAME_DVR_WAKE = TASK_SCHEDULER_FOLDER + "\\DVR wake";
         private const string TASK_NAME_DVR_SYNC = TASK_SCHEDULER_FOLDER + "\\DVR sync";
         private const string TASK_NAME_DVR_MONITOR = TASK_SCHEDULER_FOLDER + "\\DVR monitor";
+        private const string TASK_NAME_VERSION_CHECK = TASK_SCHEDULER_FOLDER + "\\Version check";
 
-        private string _dotNetExeFullPath;
-        private readonly string _workingDirectory;
-        private readonly string _dllName;
         private readonly string _libraryDatabaseFileName;
 
         public TaskScheduler()
         {
-            var fullPath = typeof(TaskScheduler).Assembly.Location;
-            _workingDirectory = Path.GetDirectoryName(fullPath);
-            _dllName = Path.GetFileName(fullPath);
             _libraryDatabaseFileName = Settings.LibraryDatabaseIsOverridden ? Settings.LibraryDatabaseFileName : null;
-        }
-
-        public bool FindDotNetExeLocation()
-        {
-            // Check the default location
-            var defaultPath = Path.Combine(Environment.GetEnvironmentVariable("ProgramFiles"), "dotnet", "dotnet.exe");
-            if (File.Exists(defaultPath))
-            {
-                Logger.LogInformation($"Found dotnet.exe at: {defaultPath}");
-                _dotNetExeFullPath = $"\"{defaultPath}\"";
-                return true;
-            }
-
-            // Check the PATH environment variable locations
-            var pathEnvVar = Environment.GetEnvironmentVariable("PATH");
-            if (!string.IsNullOrWhiteSpace(pathEnvVar))
-            {
-                var paths = pathEnvVar.Split(Path.PathSeparator);
-                foreach (var path in paths)
-                {
-                    var fullPath = Path.Combine(path, "dotnet.exe");
-                    if (File.Exists(fullPath))
-                    {
-                        Logger.LogInformation($"Found dotnet.exe at: {fullPath}");
-                        _dotNetExeFullPath = $"\"{fullPath}\"";
-                        return true;
-                    }
-                }
-            }
-
-            Logger.LogError($"Unable to find dotnet.exe at \"{defaultPath}\" or in the PATH environment variable.");
-            return false;
         }
 
         public bool CreateOrUpdateWakeUpTask(DateTime startTime)
@@ -69,8 +32,6 @@ namespace PlexDvrWaker.Plex
 
         internal bool CreateOrUpdateWakeUpTask(DateTime startTime, bool showMessageToUser)
         {
-            VerifyDotNetExeWasLocated();
-
             Logger.LogInformation($"Creating/updating wakeup task: {TASK_NAME_DVR_WAKE}");
 
             var td = TaskService.Instance.NewTask();
@@ -92,9 +53,9 @@ namespace PlexDvrWaker.Plex
 
             td.Actions.Add(new ExecAction()
             {
-                Path = _dotNetExeFullPath,
-                WorkingDirectory = _workingDirectory,
-                Arguments = _dllName + " " + Parser.Default.FormatCommandLine(
+                Path = Program.APP_EXE,
+                WorkingDirectory = Program.APP_WORKING_DIRECTORY,
+                Arguments = Parser.Default.FormatCommandLine(
                     new AddTaskOptions
                     {
                         // Recreate/update the wakeup task "after" the current recording has started.
@@ -103,16 +64,12 @@ namespace PlexDvrWaker.Plex
                         LibraryDatabaseFileName = _libraryDatabaseFileName,
                         TaskName = TASK_NAME_DVR_WAKE
                     },
-                    settings =>
-                    {
-                        settings.UseEqualToken = true;
-                        settings.ShowHidden = true;
-                    }
+                    Program.ConfigureUnParserSettings
                 )
             });
 
             var successMessage = $"Wakeup task scheduled for {trigger.StartBoundary}";
-            if (!TryCreateAdminTask(TASK_NAME_DVR_WAKE, td, successMessage, showMessageToUser))
+            if (!TryCreateTask(TASK_NAME_DVR_WAKE, td, successMessage, showMessageToUser))
             {
                 return false;
             }
@@ -122,8 +79,6 @@ namespace PlexDvrWaker.Plex
 
         public bool CreateOrUpdateDVRSyncTask(int intervalMinutes)
         {
-            VerifyDotNetExeWasLocated();
-
             Logger.LogInformation($"Creating/updating DVR sync task: {TASK_NAME_DVR_SYNC}");
 
             var td = TaskService.Instance.NewTask();
@@ -136,7 +91,7 @@ namespace PlexDvrWaker.Plex
             td.Settings.StopIfGoingOnBatteries = false;
             td.Settings.WakeToRun = false;
 
-            var tt = new TimeTrigger(new DateTime(2019, 1, 1))
+            var tt = new TimeTrigger(DateTime.Today)
             {
                 Repetition = new RepetitionPattern(TimeSpan.FromMinutes(intervalMinutes), TimeSpan.Zero)
             };
@@ -144,25 +99,21 @@ namespace PlexDvrWaker.Plex
 
             td.Actions.Add(new ExecAction()
             {
-                Path = _dotNetExeFullPath,
-                WorkingDirectory = _workingDirectory,
-                Arguments = _dllName + " " + Parser.Default.FormatCommandLine(
+                Path = Program.APP_EXE,
+                WorkingDirectory = Program.APP_WORKING_DIRECTORY,
+                Arguments = Parser.Default.FormatCommandLine(
                     new AddTaskOptions
                     {
                         Wakeup = true,
                         LibraryDatabaseFileName = _libraryDatabaseFileName,
                         TaskName = TASK_NAME_DVR_SYNC
                     },
-                    settings =>
-                    {
-                        settings.UseEqualToken = true;
-                        settings.ShowHidden = true;
-                    }
+                    Program.ConfigureUnParserSettings
                 )
             });
 
             var successMessage = $"DVR sync task scheduled for every {intervalMinutes} minute{(intervalMinutes > 1 ? "s" : "")}";
-            if (!TryCreateAdminTask(TASK_NAME_DVR_SYNC, td, successMessage, true))
+            if (!TryCreateTask(TASK_NAME_DVR_SYNC, td, successMessage, true))
             {
                 return false;
             }
@@ -172,8 +123,6 @@ namespace PlexDvrWaker.Plex
 
         public bool CreateOrUpdateDVRMonitorTask(int debounceSeconds)
         {
-            VerifyDotNetExeWasLocated();
-
             Logger.LogInformation($"Creating/updating DVR monitor task: {TASK_NAME_DVR_MONITOR}");
 
             var td = TaskService.Instance.NewTask();
@@ -197,9 +146,9 @@ namespace PlexDvrWaker.Plex
 
             td.Actions.Add(new ExecAction()
             {
-                Path = _dotNetExeFullPath,
-                WorkingDirectory = _workingDirectory,
-                Arguments = _dllName + " " + Parser.Default.FormatCommandLine(
+                Path = Program.APP_EXE,
+                WorkingDirectory = Program.APP_WORKING_DIRECTORY,
+                Arguments = Parser.Default.FormatCommandLine(
                     new MonitorOptions
                     {
                         DebounceSeconds = debounceSeconds,
@@ -207,16 +156,12 @@ namespace PlexDvrWaker.Plex
                         LibraryDatabaseFileName = _libraryDatabaseFileName,
                         TaskName = TASK_NAME_DVR_MONITOR
                     },
-                    settings =>
-                    {
-                        settings.UseEqualToken = true;
-                        settings.ShowHidden = true;
-                    }
+                    Program.ConfigureUnParserSettings
                 )
             });
 
             var successMessage = $"DVR monitor task scheduled to run at startup";
-            if (!TryCreateAdminTask(TASK_NAME_DVR_MONITOR, td, successMessage, true))
+            if (!TryCreateTask(TASK_NAME_DVR_MONITOR, td, successMessage, true))
             {
                 return false;
             }
@@ -228,16 +173,51 @@ namespace PlexDvrWaker.Plex
             return true;
         }
 
-        private bool TryCreateAdminTask(string taskPathAndName, TaskDefinition td, string successMessage, bool showMessageToUser)
+        public bool CreateOrUpdateVersionCheckTask(int intervalDays)
         {
-            // Stop the task first so that we can overwrite it
-            var task = TaskService.Instance.GetTask(taskPathAndName);
-            if (task != null && task.State == TaskState.Running)
+            Logger.LogInformation($"Creating/updating version check task: {TASK_NAME_VERSION_CHECK}");
+
+            var td = TaskService.Instance.NewTask();
+            td.RegistrationInfo.Description = $"This task will check for a newer version of {Program.APP_FRIENDLY_NAME}.";
+            td.Principal.LogonType = TaskLogonType.InteractiveToken;
+            td.Settings.AllowDemandStart = true;
+            td.Settings.DisallowStartIfOnBatteries = false;
+            td.Settings.StopIfGoingOnBatteries = false;
+            td.Settings.WakeToRun = false;
+            td.Settings.RunOnlyIfNetworkAvailable = true;
+            td.Settings.StartWhenAvailable = true;
+
+            var tt = new TimeTrigger(DateTime.Now)
             {
-                Logger.LogInformation($"  Stopping the currently running task so that it can be updated");
-                task.Stop();
+                Repetition = new RepetitionPattern(TimeSpan.FromDays(intervalDays), TimeSpan.Zero)
+            };
+            td.Triggers.Add(tt);
+
+            td.Actions.Add(new ExecAction()
+            {
+                Path = Program.APP_EXE,
+                WorkingDirectory = Program.APP_WORKING_DIRECTORY,
+                Arguments = Parser.Default.FormatCommandLine(
+                    new VersionCheckOptions
+                    {
+                        NonInteractive = true,
+                        TaskName = TASK_NAME_VERSION_CHECK
+                    },
+                    Program.ConfigureUnParserSettings
+                )
+            });
+
+            var successMessage = $"Version check task scheduled for every {intervalDays} day{(intervalDays > 1 ? "s" : "")}";
+            if (!TryCreateTask(TASK_NAME_VERSION_CHECK, td, successMessage, true))
+            {
+                return false;
             }
 
+            return true;
+        }
+
+        private static bool TryCreateTask(string taskPathAndName, TaskDefinition td, string successMessage, bool showMessageToUser)
+        {
             try
             {
                 Logger.LogInformation("  Creating/updating the task");
@@ -251,14 +231,6 @@ namespace PlexDvrWaker.Plex
             }
 
             return true;
-        }
-
-        private void VerifyDotNetExeWasLocated()
-        {
-            if (string.IsNullOrWhiteSpace(_dotNetExeFullPath))
-            {
-                throw new InvalidOperationException("dotnet.exe must be located before calling this method.");
-            }
         }
     }
 }